src/main/java/com/workshop/users/services/product/ProductServiceImpl.java
package com.workshop.users.services.product;

import com.workshop.users.api.dto.Product;
import com.workshop.users.exceptions.NotFoundProductException;
import com.workshop.users.repositories.ProductRepository;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@AllArgsConstructor
public class ProductServiceImpl implements ProductService{
    private final ProductRepository productRepository;


    public List<Product> findProductsByIds(List<Long> ids) throws NotFoundProductException {
        return productRepository.findProductsByIds(ids);
    }
}
